something remains that I couldn't solve

+++ src/main/java/com/springbootdhan/dao/CardDao.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.springbootdhan.entity.Card;
 
-public interface CardDao extends JpaRepository<Card, Integer>{}
+public interface CardDao extends JpaRepository<Card, Long>{}
